refactor(lib): improve error message handling in `getApiErrorMessages`
- Enhanced structure and readability of API error message handling.
- Refactored `getApiErrorMessages.ts` to use helper functions and constants for better maintainability and extensibility.
- Updated references to `getApiErrorMessages.ts` after renaming the file for consistency in naming conventions.
